Update onSubmit logic
