+++ src/components/Form/Form.tsx
@@ -1,14 +1,22 @@
 import { useEffect, useState } from 'react';
-import { useForm } from 'react-hook-form';
+import { SubmitHandler, useForm } from 'react-hook-form';
+import toast from 'react-hot-toast';
 import { css } from '@emotion/react';
 import { useRouter } from 'next/router';
 
-import * as S from './Form.styles';
-import useCompanyList from 'hooks/api/company/use-company-list';
-
-import { positionOptionList, onSubmit, InputListType } from './container';
 import { Modal } from 'components/common/Modal';
 import { CompanyForm } from 'components/CompanyForm';
+import useCompanyList from 'hooks/api/company/use-company-list';
+import { WorkerReqData } from 'types/worker.type';
+import axiosClient from 'libs/axios/axiosClient';
+
+import * as S from './Form.styles';
+import {
+  positionOptionList,
+  InputListType,
+  KeyListType,
+  keyList,
+} from './container';
 
 export const FormComponent: React.FC = () => {
   const { register, handleSubmit } = useForm<InputListType>();
@@ -22,6 +30,42 @@ export const FormComponent: React.FC = () => {
     router.replace(router.pathname);
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
+
+  const onSubmit: SubmitHandler<InputListType> = async data => {
+    const entries = Object.entries(data);
+
+    const allNotFilled = entries.some(([key, value]) => {
+      if (!value) {
+        toast.error(
+          `${keyList[key as keyof KeyListType]}은(는) 필수로 입력해야 해요`,
+        );
+        return true;
+      }
+    });
+
+    if (!allNotFilled) {
+      // TODO: post 로직 고도화
+      const reqData: WorkerReqData = {
+        email: data.email,
+        name: data.name,
+        worker: {
+          companyId: parseInt(data.companyId),
+          devYear: parseInt(data.devYear),
+          introduction: data.introduction,
+          position: data.position,
+        },
+      };
+
+      axiosClient
+        .post('/auth/registration', reqData)
+        .then(function (response) {
+          toast.success('회원가입이 완료되었어요');
+        })
+        .catch(function (error) {
+          toast.error(error);
+        });
+    }
+  };
 
   return (
     <S.FormWrapper>
